% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasim_SaveData.R
\name{datasim_SaveData}
\alias{datasim_SaveData}
\title{Save workflow results into the in-memory datasim data store}
\usage{
datasim_SaveData(lWorkflow, lConfig)
}
\arguments{
\item{lWorkflow}{List. Workflow object with $meta and $lResult.}

\item{lConfig}{List. Config object with $data_store.}
}
\value{
NULL (invisible). Side effect: populates lConfig$data_store.
}
\description{
Save workflow results into the in-memory datasim data store
}
