% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_LoadData.R
\name{.fs_is_absolute_path}
\alias{.fs_is_absolute_path}
\title{Test whether a path is absolute}
\usage{
.fs_is_absolute_path(path)
}
\arguments{
\item{path}{Character scalar path.}
}
\value{
Logical.
}
\description{
Test whether a path is absolute
}
\keyword{internal}
