% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/og_run.R
\name{.og_check_run_preconditions}
\alias{.og_check_run_preconditions}
\title{Validate that a project folder is runnable}
\usage{
.og_check_run_preconditions(paths, steps)
}
\description{
Validate that a project folder is runnable
}
\keyword{internal}
