% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/og_run.R
\name{.og_copy_site_page}
\alias{.og_copy_site_page}
\title{Copy the packaged site SPA to the project root as index.html (if bundled)}
\usage{
.og_copy_site_page(paths)
}
\description{
Copy the packaged site SPA to the project root as index.html (if bundled)
}
\keyword{internal}
