% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/og_init.R
\name{.og_copy_workflows}
\alias{.og_copy_workflows}
\title{Copy the canonical workflow set out of the installed gsm packages.}
\usage{
.og_copy_workflows(workflows_dir)
}
\description{
Copy the canonical workflow set out of the installed gsm packages.
}
\keyword{internal}
