% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/og_init.R
\name{.og_example_domain}
\alias{.og_example_domain}
\title{Build one example domain data.frame from gsm.core::lSource.}
\usage{
.og_example_domain(domain)
}
\description{
Build one example domain data.frame from gsm.core::lSource.
}
\keyword{internal}
