% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/og_init.R
\name{.og_rename_cols}
\alias{.og_rename_cols}
\title{Rename data.frame columns (old name -> new name), skipping absent columns.}
\usage{
.og_rename_cols(df, mapping)
}
\description{
Rename data.frame columns (old name -> new name), skipping absent columns.
}
\keyword{internal}
