% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/og_run.R
\name{.og_save_metric_outputs}
\alias{.og_save_metric_outputs}
\title{Save analyzed metric results to output/2_metrics/{id}/{step_output}.csv}
\usage{
.og_save_metric_outputs(analyzed, paths)
}
\description{
Save analyzed metric results to output/2_metrics/{id}/{step_output}.csv
}
\keyword{internal}
