% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_api.R
\name{gh_list_directory}
\alias{gh_list_directory}
\title{List directory contents via GitHub Contents API}
\usage{
gh_list_directory(repo, path, branch, token)
}
\arguments{
\item{repo}{Character. GitHub repo in "owner/repo" format.}

\item{path}{Character. Directory path within the repo.}

\item{branch}{Character. Branch name.}

\item{token}{Character. GitHub PAT.}
}
\value{
List of file/directory entry lists.
}
\description{
List directory contents via GitHub Contents API
}
